apps/syllabus/pyproject.toml
[project]
name = "syllabus"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]